# gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a chatting API",
        "title": "Jackey Chatting API",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "version": "1"
    },
    "host": "localhost:8888",
    "basePath": "/api/",
    "paths": {
...
        "/groupMessage": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get messages from a group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GroupMessage"
                ],
                "summary": "Get group messages",
                "parameters": [
                    {
                        "description": "Group ID, limit, and offset",
                        "name": "json_object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/groupMessage.GetGroupMessagesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler._BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/groupMessage.GroupMessage"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
        "/groupMessage/addGroupMember": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a member to a group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GroupMessage"
                ],
                "summary": "Add a member to a group",
                "parameters": [
                    {
                        "description": "Group ID and member ID",
                        "name": "json_object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/groupMessage.AddGroupMemberRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler._BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
        "/groupMessage/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a group with a list of members",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GroupMessage"
                ],
                "summary": "Create a message group",
                "parameters": [
                    {
                        "description": "Group name and member IDs",
                        "name": "json_object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/groupMessage.CreateGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler._BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
        "/groupMessage/send": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send a message to a group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GroupMessage"
                ],
                "summary": "Send a group message",
                "parameters": [
                    {
                        "description": "Group ID and content",
                        "name": "json_object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/groupMessage.SendGroupMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler._BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "groupMessage.AddGroupMemberRequest": {
            "type": "object",
            "properties": {
                "group_id": {
                    "description": "群组ID",
                    "type": "integer"
                },
                "member_id": {
                    "description": "被添加的用户ID",
                    "type": "integer"
                },
                "user_id": {
                    "description": "发起请求的用户ID",
                    "type": "integer"
                }
            }
        },
        "groupMessage.CreateGroupRequest": {
            "type": "object",
            "properties": {
                "group_name": {
                    "description": "群组名称",
                    "type": "string"
                },
                "user_id": {
                    "description": "创建群组的用户ID",
                    "type": "integer"
                }
            }
        },
        "groupMessage.GetGroupMessagesRequest": {
            "type": "object",
            "properties": {
                "group_id": {
                    "description": "群组ID",
                    "type": "integer"
                },
                "limit": {
                    "description": "每次请求返回的最大消息数",
                    "type": "integer"
                },
                "offset": {
                    "description": "从第几条记录开始读取",
                    "type": "integer"
                },
                "user_id": {
                    "description": "发起请求的用户ID",
                    "type": "integer"
                }
            }
        },
        "groupMessage.GroupMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "group_id": {
                    "type": "integer"
                },
                "msg_id": {
                    "type": "integer"
                },
                "sender_id": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "groupMessage.SendGroupMessageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "group_id": {
                    "type": "integer"
                },
                "sender_id": {
                    "type": "integer"
                }
            }
        },
        "handler._BaseResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "数据"
                },
                "status": {
                    "description": "状态",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "handler._ResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误信息",
                    "type": "string",
                    "example": "err msg"
                },
                "status": {
                    "description": "状态",
                    "type": "string",
                    "example": "failed"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}